requirements.txt
asgiref==3.7.2
click==8.1.5
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h11==0.14.0
httpagentparser==1.9.5
mysqlclient==2.2.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
uvicorn==0.23.0